Language.FineTypes.Export.OpenAPI.Schema

Description

Export type definitions to OpenAPI JSON schemas.

https://www.openapis.org

Documentation

schemaFromModule :: Module -> OpenApi

Export

Assumes that the argument satisfies supportsJSON.

schemaFromTyp :: Typ -> Schema

supportsJSON :: Module -> Bool

Test whether a Module only uses types supported by JSON.

JSON does not support finite maps such as ↦, ↦0, →∗.

convertToJSON :: Declarations -> Declarations

Convert Typ definitions to JSON.

The result satisfies supportsJSON.

Note: We don't recommend that you use this function, because it does a lot of conversions under the hood. Instead, if you want to export a Typ to JSON, we recommend that you explicitly define a second Typ which is apparently compatible with JSON, and show that the first Typ can be embedded into the second Typ.