Language.FineTypes.Module

Description

A Module is a collection of Typ definitions.

Documentation

type ModuleName = String

data Module

A Module is a collection of Typ definitions and documentation.

Constructors

Module

Fields

moduleName :: ModuleName

moduleImports :: Imports

moduleDeclarations :: Declarations

newtype Import

Constructors

ImportNames

Fields

getImportNames :: Set TypName

type Imports = Map ModuleName Import

type Declarations = Map TypName Typ

resolveImports :: Map ModuleName Module -> Module -> Maybe Declarations

For a given Module, resolve import by adding the declarations for the imported names.

Fails if an imported name cannot be found.

collectNotInScope :: Module -> Set TypName

Collect all Typ names that have not been defined or imported in the Module.

resolveVars :: Declarations -> Typ -> Typ

Resolve all variables in a Typ that can be resolved using the given declarations.

This function will loop in the case of recursive definitions.