Language.FineTypes.Typ

Description

Typ represents the types that can be defined with FineTypes.

Typ

type TypName = String

type ConstructorName = String

type FieldName = String

type VarName = String

data Typ

A Typ describes a set of values.

We call it Typ, because Type is reserved for Haskell types.

You can obtain new types from old types with mathematical constructions, such as taking disjoint sums or cartesian products.

Constructors

Abstract
Typ which is unspecified.

Var TypName
Variable

Zero TypConst
Known Typ constant.

One OpOne Typ
Apply an unary operation to a Typ.

Two OpTwo Typ Typ
Apply a binary operation to two Typs.

ProductN [(FieldName, Typ)]
Cartesian product with component names.

SumN [(ConstructorName, Typ)]
Disjoint union with constructor names.

Constrained VarName Typ Constraint
A type with a value constraint

data TypConst

Predefined Typ.

Constructors

Bool
Booleans, Bool.

Bytes
Sequences of bytes, Bytes.

Integer
Integers, ℤ.

Natural
Natural numbers, ℕ.

Text
A sequence of unicode characters, Text.

Unit
Type with a single element, Unit.

Rational
Rational numbers, ℚ.

data OpOne

Unary operations on Typ.

Constructors

Option
Given a set A, A? is the set of elements of A and a separate element (often called "null").

Sequence
Given a set A, A* is the set of sequences having elements taken from A.

PowerSet
Given a set A, ℙ A is the set of all the subsets of A.

data OpTwo

Binary operations on Typ.

Constructors

Sum2
A ⊎ B denotes the disjoint union of A and B.

Product2
A × B denotes the cartesian product of A and B.

PartialFunction
A ↦ B denotes a partial function from A to B, which can be seen as a map (dictionary) with keys in A and values in B.

FiniteSupport
A →∗ B denotes a finitely supported partial function.

data Constraint1

Constructors

Braces Constraint

Token String
String contains anything but whitespace and curly braces

type Constraint = [Constraint1]

Traversals

everywhere :: (Typ -> Typ) -> Typ -> Typ

Apply a transformation everywhere; bottom-up.

everything :: (r -> r -> r) -> (Typ -> r) -> Typ -> r

Summarise all nodes; top-down, left-to-right.

depth :: Typ -> Int