Language.FineTypes.Typ.Gen

Documentation

data Mode

If the generated Typ should be concrete or not. Concrete will not contain Abstract or Var leaves

Constructors

Complete

Concrete

onComplete :: Monoid p => Mode -> p -> p

type DepthGen = Int

Minimum depth of the generated Typ. Shorter than depth branches are still possible if the actual Typ is Zero or Abstract or Var

data Branch

Whether the generated Typ will be Zero or more complex

Constructors

Wont

Will

willBranch :: DepthGen -> Gen Branch

onWill :: Monoid p => Branch -> p -> p

data WithConstraints

Whether the generated Typ will have constraints or not

Constructors

WithConstraints

WithoutConstraints

onConstraints :: Monoid p => WithConstraints -> p -> p

data Round

Whether the generated Typ is the top one in the tree or not

Constructors

Top

Rest

onTop :: Monoid p => Round -> p -> p

genTypFiltered

Arguments

:: (Typ -> Bool)
Whether the generated Typ should be filtered out

-> WithConstraints
Whether the generated Typ can have constraints or not

-> Mode
Whether the generated Typ should be concrete or not

-> DepthGen
Maximum depth of the generated Typ

-> Gen Typ

Generate a random Typ.

genTyp :: WithConstraints -> Mode -> DepthGen -> Gen Typ

genConstraint :: Int -> Gen Constraint

genConstraint1 :: Int -> Gen Constraint1

goodToken :: String -> Bool

goodChar :: Char -> Bool

genConstrainedTyp :: Mode -> Gen Typ

genTagged :: Gen [a] -> Gen Typ -> Gen [(a, Typ)]

genTwoOpen :: Gen OpTwo

genTwoClose :: Gen OpTwo

genOne :: Gen OpOne

genConst :: Gen TypConst

notPrivate :: String -> Bool

genNames :: Gen [String]

genName :: Gen [Char]

genConstructors :: Gen [ConstructorName]

genVarName :: Gen TypName

capitalise :: [Char] -> [Char]

genFields :: Gen [FieldName]

logScale :: Double -> Gen a -> Gen a

shrinkTyp :: Typ -> [Typ]

shrinkNamed :: (t, Typ) -> [(t, Typ)]

shrinkConstraint :: Constraint -> [Constraint]

shrinkConstraint1 :: Constraint1 -> [Constraint1]