Language.FineTypes.Export.Haskell.Value.Runtime

Description

Export values of Haskell types to Value.

Documentation

class (Ord a, ToTyp a) => ToValue a where

Class of types that can be converted to Value.

Note: The Ord constraint is necessary to deal with Set.

Methods

toValue :: a -> Value

data Value

Constructors

Zero ZeroF

One (OneF Value)

Two (TwoF Value Value)

Product [Value]
N-ary products.

Sum Ix Value
N-ary sums.

class ToTyp a where

Methods

toTyp :: Proxy a -> Typ