Language.FineTypes.Module

Description

A Module is a collection of Identifier definitions.

Documentation

type ModuleName = String

data Module

A Module is a collection of Identifier definitions and documentation.

Constructors

Module

Fields

moduleName :: ModuleName

moduleImports :: Imports

moduleDeclarations :: Declarations

moduleDocumentation :: Documentation

newtype Import

Constructors

ImportNames

Fields

getImportNames :: Set TypName

type Imports = Map ModuleName Import

type Declarations = Map TypName Typ

Documentation texts

data Identifier

An Identifier refers to a specific TypName or one of its fields or constructors.

Constructors

Typ TypName

Field TypName FieldName

Constructor TypName ConstructorName

type DocString = String

newtype Documentation

Documentation texts associated to Identifiers.

Constructors

Documentation

Fields

getDocumentation :: Map Identifier (Map Place DocString)

data Place

Place for putting documentation

Constructors

BeforeMultiline

Before

After

document :: Identifier -> Map Place DocString -> Documentation

Name resolution

resolveImports :: Map ModuleName Module -> Module -> Maybe Declarations

For a given Module, resolve import by adding the declarations for the imported names.

Fails if an imported name cannot be found.

collectNotInScope :: Module -> Set TypName

Collect all Identifier names that have not been defined or imported in the Module.

resolveVars :: Declarations -> Typ -> Typ

Resolve all variables in a Identifier that can be resolved using the given declarations.

This function will loop in the case of recursive definitions.