Language.FineTypes.Package

Description

Convenient packaging of Package and PackageDescription functionality.

Package

newtype Package

A package is a collection of module or signature definitions.

Constructors

Package

Fields

packageModules :: Map ModuleName (Either Signature Module)

emptyPackage :: Package

includePackage :: Package -> Package -> Either ErrIncludePackage Package

Include a package in the current package if possible.

addModule :: Module -> Package -> Either ErrAddModule Package

Add a module to the current package if possible.

Package descriptions

data PackageDescription

A PackageDescription is a sequence of Statement that evaluate to a Package.

Constructors

PackageDescription

Fields

packageName :: PackageName

packageStatements :: [Statement]

type PackageName = String

data Statement

Constructors

Include PackageName Source

Module ModuleName Source

Signature ModuleName Source

Assert Assertion

newtype Source

Constructors

File FilePath

parsePackageDescription :: String -> Either ErrParsePackage PackageDescription

compilePackageDescription :: FilePath -> PackageDescription -> IO (Either ErrCompilePackage Package)

Compile a package description to a Package or return a descriptive error message.

The first argument is a directory which is used as base for resolving relative filenames.

Error types

newtype ErrParsePackage

Constructors

ErrParsePackage

Fields

parseErrorBundle :: ParseErrorBundle String Void

data ErrCompilePackage

Constructors

ErrFile IOError
Error while reading a file.

ErrParseModuleError ModuleName ErrParseModule
Error while parsing a module.

ErrIncludeParsePackageError PackageName ErrParsePackage
Error while parsing an included package description.

ErrIncludeCompilePackage PackageName ErrCompilePackage
Error while compiling an included package.

ErrIncludePackage PackageName ErrIncludePackage
Error while including a package.

ErrIncludePackageNameMismatch PackageName PackageName
The package name in the include statement does not match the included package name

ErrAddModule ModuleName ErrAddModule
Error while adding a module.

ErrAddModuleNameMismatch ModuleName ModuleName
The module name in the module statement does not match the included module name

newtype ErrIncludePackage

Constructors

ErrModulesAlreadyInScope (Set ModuleName)

data ErrAddModule

Constructors

ErrModuleAlreadyInScope
The module is already in scope.

ErrImportNotInScope (Set (ModuleName, TypName))
The module imports a type that is not exported.

ErrNamesNotInScope (Set TypName)
The added module uses a type which is not defined.