Language.FineTypes.Typ.Gen

Documentation

data Concrete

If the generated Typ should be concrete or not. Concrete will not contain Abstract or Var leaves

Constructors

Complete

Concrete

patchNoData :: Concrete -> [Gen Typ] -> [Gen Typ]

type DepthGen = Int

Minimum depth of the generated Typ. Shorter than depth branches are still possible if the actual Typ is Zero or Abstract or Var

willBranch :: DepthGen -> Gen Bool

genTyp :: Concrete -> DepthGen -> Gen Typ

Generate a random Typ.

genTwo :: Gen OpTwo

genOne :: Gen OpOne

genConst :: Gen TypConst

genNames :: Gen [String]

genName :: Gen [Char]

genConstructors :: Gen [ConstructorName]

genVarName :: Gen TypName

capitalise :: [Char] -> [Char]

genFields :: Gen [FieldName]

logScale :: Double -> Gen a -> Gen a